refactor: cleanup swkbd code
