+++ src/ship/CMakeLists.txt
@@ -104,6 +104,8 @@ elseif (CMAKE_SYSTEM_NAME STREQUAL "NintendoSwitch")
         ${CMAKE_CURRENT_SOURCE_DIR}/port/switch/SwitchController.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/port/switch/SwitchImpl.h
         ${CMAKE_CURRENT_SOURCE_DIR}/port/switch/SwitchImpl.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/port/switch/SwitchKeyboard.h
+        ${CMAKE_CURRENT_SOURCE_DIR}/port/switch/SwitchKeyboard.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/port/switch/SwitchPerformanceProfiles.h
     )
 endif()
