+++ src/ship/port/switch/SwitchImpl.cpp
@@ -1,5 +1,6 @@
 #ifdef __SWITCH__
 #include "SwitchImpl.h"
+#include "SwitchKeyboard.h"
 #include <switch.h>
 #include <SDL2/SDL.h>
 #include "SwitchPerformanceProfiles.h"
@@ -9,30 +10,12 @@
 #include "ship/audio/Audio.h"
 #include "ship/utils/StringHelper.h"
 
-#include <imgui_internal.h>
-
-namespace ImStb {
-#include <imstb_textedit.h>
-}
-
 #define DOCKED_MODE 1
 #define HANDHELD_MODE 0
 
 static AppletHookCookie applet_hook_cookie;
 static bool isRunning = true;
 static bool hasFocus = true;
-
-static SwkbdInline sKeyboard;
-static bool sIsKeyboardLaunched = false;
-static bool sIsKeyboardActive = false;
-static std::string sKeyboardText;
-static bool sIsKeyboardTextChanged = false;
-static bool sIsKeyboardSubmitted = false;
-static ImGuiID sKeyboardOwner = 0;
-static int sKeyboardCursorPos = -1;
-static bool sIsCursorMoved = false;
-static float sKeyboardYOffset = 0.0f;
-static int sPendingCursorPos = -1;
 
 HidsysUniquePadId uniquePadIds[8];
 
@@ -66,11 +49,7 @@ void Ship::Switch::Init(SwitchPhase phase) {
 }
 
 void Ship::Switch::Exit() {
-    if (sIsKeyboardLaunched) {
-        swkbdInlineClose(&sKeyboard);
-        sIsKeyboardLaunched = false;
-    }
-
+    Keyboard::Close();
     socketExit();
     clkrstExit();
     appletSetGamePlayRecordingState(false);
@@ -102,231 +81,6 @@ void Ship::Switch::ImGuiSetupFont(ImFontAtlas* fonts) {
     fonts->Build();
 
     plExit();
-}
-
-static void OnKeyboardStringChanged(const char* str, SwkbdChangedStringArg* arg) {
-    sKeyboardText = str ? str : "";
-    sIsKeyboardTextChanged = true;
-    sKeyboardCursorPos = arg->cursorPos;
-}
-
-static void OnKeyboardMovedCursor(const char* str, SwkbdMovedCursorArg* arg) {
-    sKeyboardCursorPos = arg->cursorPos;
-    sIsCursorMoved = true;
-}
-
-static void OnKeyboardDecidedEnter(const char* str, SwkbdDecidedEnterArg*) {
-    sKeyboardText = str ? str : "";
-    sIsKeyboardTextChanged = true;
-    sIsKeyboardSubmitted = true;
-    swkbdInlineDisappear(&sKeyboard);
-    sIsKeyboardActive = false;
-}
-
-static void OnKeyboardDecidedCancel() {
-    // Keep whatever the player left in the field -- including empty, after erasing everything.
-    // The cancel button already disappears the keyboard, so don't call Disappear here (it corrupts the applet and
-    // breaks the next Appear).
-    sIsKeyboardTextChanged = true;
-    sIsKeyboardActive = false;
-}
-
-void Ship::Switch::CreateKeyboard() {
-    auto result = swkbdInlineCreate(&sKeyboard);
-    if (R_FAILED(result)) {
-        SPDLOG_ERROR("swkbdInlineCreate failed: {:#x}", result);
-        return;
-    }
-
-    result = swkbdInlineLaunchForLibraryApplet(&sKeyboard, SwkbdInlineMode_AppletDisplay, 0);
-    if (R_FAILED(result)) {
-        SPDLOG_ERROR("swkbdInlineLaunchForLibraryApplet failed: {:#x}", result);
-        swkbdInlineClose(&sKeyboard);
-        return;
-    }
-
-    swkbdInlineSetChangedStringCallback(&sKeyboard, OnKeyboardStringChanged);
-    swkbdInlineSetMovedCursorCallback(&sKeyboard, OnKeyboardMovedCursor);
-    swkbdInlineSetDecidedEnterCallback(&sKeyboard, OnKeyboardDecidedEnter);
-    swkbdInlineSetDecidedCancelCallback(&sKeyboard, OnKeyboardDecidedCancel);
-
-    sIsKeyboardLaunched = true;
-}
-
-// The inline keyboard is event-driven: this pumps applet replies and fires the callbacks.
-// It must run every frame once launched.
-void Ship::Switch::UpdateKeyboard() {
-    if (!sIsKeyboardLaunched) {
-        return;
-    }
-
-    swkbdInlineUpdate(&sKeyboard, nullptr);
-}
-
-void Ship::Switch::ImGuiProcessEvent(bool wantsTextInput) {
-}
-
-static bool sPendingCancel = false;
-
-static void SetInputTextCursorPos(ImGuiInputTextState* state, int pos) {
-    if (!state || !state->Stb) {
-        return;
-    }
-    pos = ImClamp(pos, 0, state->TextLen);
-    state->Stb->cursor = pos;
-    state->Stb->select_start = pos;
-    state->Stb->select_end = pos;
-    state->CursorFollow = true;
-}
-
-void Ship::Switch::HandleInlineKeyboard() {
-    if (!sIsKeyboardLaunched) {
-        return;
-    }
-
-    ImGuiContext& g = *GImGui;
-    ImGuiIO& io = ImGui::GetIO();
-
-    if (sPendingCancel) {
-        sPendingCancel = false;
-        if (g.ActiveId == sKeyboardOwner) {
-            ImGui::ClearActiveID();
-        }
-        sKeyboardOwner = 0;
-        sPendingCursorPos = -1;
-    }
-
-    if (sPendingCursorPos >= 0) {
-        ImGuiInputTextState* state = ImGui::GetInputTextState(sKeyboardOwner);
-        if (state) {
-            SetInputTextCursorPos(state, sPendingCursorPos);
-        }
-        sPendingCursorPos = -1;
-    }
-
-    if (io.WantTextInput && !sIsKeyboardActive && sKeyboardOwner == 0) {
-        ImGuiInputTextState* state = ImGui::GetInputTextState(g.ActiveId);
-        if (state) {
-            std::string initialText;
-            if (state->TextLen > 0) {
-                initialText.assign(state->TextA.Data, state->TextLen);
-            }
-            ShowKeyboard(g.ActiveId, initialText);
-        }
-    }
-
-    if (sIsCursorMoved && !sIsKeyboardTextChanged) {
-        ImGuiInputTextState* state = ImGui::GetInputTextState(sKeyboardOwner);
-        if (state && sKeyboardCursorPos >= 0) {
-            SetInputTextCursorPos(state, sKeyboardCursorPos);
-        }
-        sIsCursorMoved = false;
-    }
-
-    if (sIsKeyboardTextChanged) {
-        sIsKeyboardTextChanged = false;
-        sIsCursorMoved = false;
-
-        ImGuiInputTextState* state = ImGui::GetInputTextState(sKeyboardOwner);
-        if (state) {
-            state->ClearText();
-            const char* s = sKeyboardText.c_str();
-            while (*s) {
-                unsigned int c;
-                int adv = ImTextCharFromUtf8(&c, s, nullptr);
-                if (c == 0) break;
-                io.InputQueueCharacters.push_back((ImWchar)c);
-                s += adv;
-            }
-            sPendingCursorPos = sKeyboardCursorPos;
-        }
-
-        if (sIsKeyboardSubmitted) {
-            sIsKeyboardSubmitted = false;
-            io.SetAppAcceptingEvents(true);
-            io.AddKeyEvent(ImGuiKey_Enter, true);
-            io.AddKeyEvent(ImGuiKey_Enter, false);
-        } else if (!sIsKeyboardActive) {
-            sPendingCancel = true;
-        }
-    }
-
-    if (!io.WantTextInput && sIsKeyboardActive) {
-        swkbdInlineDisappear(&sKeyboard);
-        sIsKeyboardActive = false;
-        sKeyboardOwner = 0;
-    }
-
-    if (!io.WantTextInput && !sIsKeyboardActive) {
-        sKeyboardOwner = 0;
-    }
-
-    if (sIsKeyboardActive && g.PlatformImeDataPrev.WantVisible) {
-        float inputY = g.PlatformImeDataPrev.InputPos.y + g.PlatformImeDataPrev.InputLineHeight;
-        float keyboardTop = io.DisplaySize.y * 0.4f;
-        float margin = 0;
-        if (inputY > keyboardTop - margin) {
-            sKeyboardYOffset = inputY - keyboardTop + margin;
-            sKeyboardYOffset = ImMin(sKeyboardYOffset, g.PlatformImeDataPrev.InputPos.y - margin);
-            sKeyboardYOffset = ImMax(sKeyboardYOffset, 0.0f);
-        } else {
-            sKeyboardYOffset = 0.0f;
-        }
-    } else {
-        sKeyboardYOffset = 0.0f;
-    }
-
-    io.SetAppAcceptingEvents(!sIsKeyboardActive);
-}
-
-float Ship::Switch::GetKeyboardYOffset() {
-    return sKeyboardYOffset;
-}
-
-
-void Ship::Switch::ShowKeyboard(ImGuiID owner, const std::string& initialText) {
-    if (!sIsKeyboardLaunched || sIsKeyboardActive) {
-        return;
-    }
-
-    sKeyboardOwner = owner;
-    sKeyboardText = initialText;
-    sIsKeyboardSubmitted = false;
-    swkbdInlineSetInputText(&sKeyboard, initialText.c_str());
-    swkbdInlineSetCursorPos(&sKeyboard, static_cast<std::int32_t>(initialText.length()));
-
-    SwkbdAppearArg appear = {};
-    swkbdInlineMakeAppearArg(&appear, SwkbdType_QWERTY);
-    swkbdInlineAppearArgSetStringLenMax(&appear, 255);
-    swkbdInlineAppear(&sKeyboard, &appear);
-
-    sIsKeyboardActive = true;
-}
-
-bool Ship::Switch::IsKeyboardActive() {
-    return sIsKeyboardActive;
-}
-
-// Returns true and fills out with the latest keyboard text exactly once per change (typing, enter, or cancel-revert).
-// One-shot so it doesn't continually overwrite the filter and fight other edits.
-bool Ship::Switch::ConsumeKeyboardText(ImGuiID owner, std::string& out, bool* isSubmitted) {
-    if (!sIsKeyboardTextChanged || sKeyboardOwner != owner) {
-        if (isSubmitted) {
-            *isSubmitted = false;
-        }
-
-        return false;
-    }
-
-    sIsKeyboardTextChanged = false;
-    out = sKeyboardText;
-
-    if (isSubmitted) {
-        *isSubmitted = sIsKeyboardSubmitted;
-    }
-
-    sIsKeyboardSubmitted = false;
-    return true;
 }
 
 bool Ship::Switch::IsRunning() {
@@ -395,11 +149,8 @@ static void on_applet_hook(AppletHookType hook, void* param) {
                 }
             } else {
                 Ship::Switch::ApplyOverclock();
-                // reinitialize audio subsystem to fix audio problems after resuming from sleep
-                // see https://github.com/HarbourMasters/Shipwright/issues/3317
                 SPDLOG_INFO("restarting SDL audio system to work around audio problems on resume");
                 if (auto audio = Ship::Context::GetRawInstance()->GetAudio(); audio != nullptr) {
-                    // the audio subsystem is not initialized during applet boot
                     audio->SetCurrentAudioBackend(Ship::AudioBackend::SDL);
                 }
             }
@@ -422,7 +173,6 @@ static void on_applet_hook(AppletHookType hook, void* param) {
 void Ship::Switch::ShowErrorApplet(const char* format, ...) {
     ErrorSystemConfig errorConfig = {};
 
-    // Error applet can display up to 2048 bytes
     char messageBuffer[2048];
     va_list args;
     va_start(args, format);
@@ -432,7 +182,7 @@ void Ship::Switch::ShowErrorApplet(const char* format, ...) {
     const Result rc = errorSystemCreate(&errorConfig, messageBuffer, nullptr);
 
     if (R_SUCCEEDED(rc)) {
-        errorSystemSetResult(&errorConfig, MAKERESULT(400, 1)); // module id, error code
+        errorSystemSetResult(&errorConfig, MAKERESULT(400, 1));
         errorSystemShow(&errorConfig);
     }
 }
